+++ public/index.html
@@ -37,6 +37,6 @@
 
       <!-- Контент -->
     </main>
-    <script src="../src/main.js"></script>
+    <script src="bundle.js"></script>
   </body>
 </html>
