+++ src/components/create-element.js
@@ -1,0 +1,8 @@
+const createElement = (template) => {
+  const parent = document.createElement(`div`);
+  parent.innerHTML = template;
+
+  return parent.firstElementChild;
+};
+
+export default createElement;
